Create necessary files and folders, and also...
...added the require(file) statements in the init.lua file.
